refactored line labels.
